fix(hw2): type warning AGAIN, and 0204 IEEE 754

+++ hw2/hw0201.c
@@ -3,16 +3,16 @@
 
 int main() {
     // ax^2 + bx + c = 0
-    int64_t a, b, c;
+    int32_t a, b, c;
     printf("Please enter a quadratic polynomial (a,b,c): ");
     // Throw an ERROR if the input is not accepted
-    if (scanf("%lli,%lli,%lli", &a, &b, &c) != 3 || a == 0) {
+    if (scanf("%d,%d,%d", &a, &b, &c) != 3 || a == 0) {
         printf("Invalid Input!\n");
         return 1;
     }
 
     // Calculate discriminant
-    int64_t d = b * b - 4 * a * c;
+    int32_t d = b * b - 4 * a * c;
 
     if (d < 0) printf("No real roots.\n");
     else if (d == 0) printf("One real root.\n");
